ollama_test.py
import requests

url = "http://localhost:11434/api/generate"

data = {
    "model": "llama3",
    "prompt": "Patient has deep leg wound and heavy bleeding. What is the first triage action?",
    "stream": False
}

response = requests.post(url, json=data)

result = response.json()

print(result["response"])