requirements.txt
sentence-transformers
faiss-cpu
scikit-learn
numpy
openai
pyttsx3
faster-whisper
torch